@@ -22,8 +22,8 @@
     "babel-core": "6.26.3",
     "immutability-helper": "2.7.0",
     "moment": "2.24.0",
-    "react": "16.7.0",
-    "react-dom": "16.7.0",
+    "react": "16.8.3",
+    "react-dom": "16.8.3",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
     "react-modal": "3.4.5",
@@ -43,7 +43,7 @@
     "@types/enzyme": "3.1.15",
     "@types/jest": "23.3.13",
     "@types/react": "16.8.3",
-    "@types/react-dom": "16.8.1",
+    "@types/react-dom": "16.8.3",
     "@types/react-modal": "3.8.1",
     "@types/react-redux": "7.0.1",
     "@types/react-router-dom": "4.3.1",
@@ -70,7 +70,7 @@
     "node-sass": "4.11.0",
     "prettier": "1.16.4",
     "prettier-tslint": "0.4.2",
-    "react-test-renderer": "16.7.0",
+    "react-test-renderer": "16.8.3",
     "react-testing-library": "5.4.4",
     "redux-mock-store": "1.5.3",
     "resolve-url-loader": "3.0.0",
