@@ -27,8 +27,8 @@
     "dot-object": "1.7.1",
     "immutability-helper": "2.7.0",
     "lodash.isequal": "4.5.0",
-    "react": "16.7.0",
-    "react-dom": "16.7.0",
+    "react": "16.8.3",
+    "react-dom": "16.8.3",
     "react-dropzone": "^9.0.0",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
@@ -46,14 +46,14 @@
     "webpack-dev-server": "3.1.14"
   },
   "devDependencies": {
-    "@types/classnames": "2.2.6",
+    "@types/classnames": "^2.2.6",
     "@types/deepmerge": "2.1.0",
     "@types/dot-object": "1.5.0",
     "@types/enzyme": "3.1.15",
     "@types/jest": "23.3.13",
     "@types/lodash.isequal": "4.5.3",
     "@types/react": "16.8.3",
-    "@types/react-dom": "16.8.1",
+    "@types/react-dom": "16.8.3",
     "@types/react-modal": "3.8.1",
     "@types/react-redux": "7.0.1",
     "@types/react-router-dom": "4.3.1",
@@ -64,7 +64,7 @@
     "awesome-typescript-loader": "5.0.0",
     "babel-jest": "23.6.0",
     "babel-preset-react": "6.24.1",
-    "classnames": "2.2.6",
+    "classnames": "^2.2.6",
     "cross-env": "5.1.6",
     "css-loader": "0.28.11",
     "enzyme": "3.8.0",
@@ -83,7 +83,7 @@
     "prettier-tslint": "0.4.2",
     "react-dnd": "7.0.2",
     "react-dnd-html5-backend": "7.0.2",
-    "react-test-renderer": "16.7.0",
+    "react-test-renderer": "16.8.3",
     "redux-mock-store": "1.5.3",
     "resolve-url-loader": "3.0.0",
     "rimraf": "2.6.2",
