@@ -21,8 +21,8 @@
     "classnames": "2.2.6",
     "immutability-helper": "2.7.0",
     "monaco-editor-webpack-plugin": "1.7.0",
-    "react": "16.7.0",
-    "react-dom": "16.7.0",
+    "react": "16.8.3",
+    "react-dom": "16.8.3",
     "react-grid-layout": "0.16.6",
     "react-hot-loader": "4.3.1",
     "react-modal": "3.4.5",
@@ -42,7 +42,7 @@
     "@types/enzyme": "3.1.15",
     "@types/jest": "23.3.13",
     "@types/react": "16.8.3",
-    "@types/react-dom": "16.8.1",
+    "@types/react-dom": "16.8.3",
     "@types/react-modal": "3.8.1",
     "@types/react-redux": "7.0.1",
     "@types/react-router-dom": "4.3.1",
@@ -68,7 +68,7 @@
     "node-sass": "4.11.0",
     "prettier": "1.16.4",
     "prettier-tslint": "0.4.2",
-    "react-test-renderer": "16.7.0",
+    "react-test-renderer": "16.8.3",
     "redux-mock-store": "1.5.3",
     "resolve-url-loader": "3.0.0",
     "rimraf": "2.6.2",
