@@ -36,9 +36,9 @@
     "jest-enzyme": "7.0.1",
     "jest-junit": "6.1.0",
     "jsdom": "13.2.0",
-    "react": "16.7.0",
-    "react-dom": "16.7.0",
-    "react-test-renderer": "16.7.0",
+    "react": "16.8.3",
+    "react-dom": "16.8.3",
+    "react-test-renderer": "16.8.3",
     "ts-jest": "23.10.5",
     "typescript": "3.1.6"
   },
