@@ -1,3 +1,4 @@
+import { shallow, mount } from 'enzyme';
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
@@ -19,12 +20,13 @@ describe('>>> components/base/CheckboxesContainerComponent.tsx --- Snapshot', ()
       id: mockId,
       title: 'test-checkboxescontainer',
       component: 'Checkboxes',
+      readOnly: false,
       options: [{
         label: 'test-label-1',
         value: 'test-1',
       }, {
-        label: 'test-label-1',
-        value: 'test-1',
+        label: 'test-label-2',
+        value: 'test-2',
       }],
     };
     mockHandleDataChange = (data: any) => null;
@@ -47,5 +49,71 @@ describe('>>> components/base/CheckboxesContainerComponent.tsx --- Snapshot', ()
       />,
     );
     expect(rendered).toMatchSnapshot();
+  });
+  it('+++ should render editable component when readOnly is false', () => {
+    const shallowCheckbox = shallow(
+      <CheckboxContainerComponent
+        id={mockId}
+        component={mockComponent}
+        formData={mockFormData}
+        handleDataChange={mockHandleDataChange}
+        getTextResource={mockGetTextResource}
+        isValid={mockIsValid}
+        designMode={mockDesignMode}
+        validationMessages={{}}
+      />,
+    );
+    expect(shallowCheckbox.find({ type: 'checkbox' })).toHaveLength(2);
+    expect(shallowCheckbox.find('.custom-control-label').first().hasClass('disabled-checkbox')).toBe(false);
+  });
+  it('+++ should render un-editable component when readOnly is true', () => {
+    const shallowCheckbox = shallow(
+      <CheckboxContainerComponent
+        id={mockId}
+        component={{
+          id: mockId,
+          component: 'Checkboxes',
+          readOnly: true,
+          options: [{
+            label: 'test-label-1',
+            value: 'test-1',
+          }, {
+            label: 'test-label-2',
+            value: 'test-2',
+          }],
+        }}
+        formData={mockFormData}
+        handleDataChange={mockHandleDataChange}
+        getTextResource={mockGetTextResource}
+        isValid={mockIsValid}
+        designMode={mockDesignMode}
+        validationMessages={{}}
+      />,
+    );
+    expect(shallowCheckbox.find('.custom-control-label').first().hasClass('disabled-checkbox')).toBe(true);
+  });
+  it('+++ checked prop should change onClick', () => {
+    const mountedCheckbox = mount(
+      <CheckboxContainerComponent
+        id={mockId}
+        component={mockComponent}
+        formData={mockFormData}
+        handleDataChange={mockHandleDataChange}
+        getTextResource={mockGetTextResource}
+        isValid={mockIsValid}
+        designMode={mockDesignMode}
+        validationMessages={{}}
+      />,
+    );
+    const instance = mountedCheckbox.instance() as CheckboxContainerComponent;
+    const checkbox = mountedCheckbox.find({ type: 'checkbox' }).first();
+    const customControl = mountedCheckbox.find('.custom-control').first();
+    expect(checkbox.props().checked).toBe(false);
+    expect(customControl.is('div')).toBe(true);
+    const spy = jest.spyOn(instance, 'onDataChanged');
+    customControl.simulate('click', { value: 'test-1' });
+    customControl.simulate('click', { value: 'test-1' });
+    expect(spy).toHaveBeenCalled();
+    expect(checkbox.props().value).toBe('test-1');
   });
 });
