@@ -1,8 +1,10 @@
 import axios, { AxiosResponse } from 'axios';
+import classNames = require('classnames');
 import update from 'immutability-helper';
 import * as React from 'react';
 
 import '../../styles/AddressComponent.css';
+import '../../styles/shared.css';
 import { renderValidationMessagesForComponent } from '../../utils/render';
 
 export interface IAddressComponentProps {
@@ -148,8 +150,8 @@ export class AddressComponent extends React.Component<IAddressComponentProps, IA
   }
 
   public joinValidationMessages = (): IComponentValidations => {
-    const {validations} = this.state as any;
-    let {validationMessages} = this.props;
+    const { validations } = this.state as any;
+    let { validationMessages } = this.props;
 
     for (const fieldKey in validations) {
       if (!validations[fieldKey]) {
@@ -196,44 +198,47 @@ export class AddressComponent extends React.Component<IAddressComponentProps, IA
             }
           </label>
           <input
-            className={'form-control'}
+            className={'form-control' + (this.props.component.readOnly ? ' disabled' : '')}
             value={address}
             onChange={this.updateField.bind(null, AddressKeys.address)}
             onBlur={this.onBlurField.bind(null, AddressKeys.address)}
+            disabled={this.props.component.readOnly}
           />
           {validations ?
             renderValidationMessagesForComponent(validations[AddressKeys.address], `${id}_${AddressKeys.address}`)
-          : null}
+            : null}
           <div className={'address-component-postplace-zipCode'}>
             <div className={'address-component-zipCode'}>
               <label className={'address-component-label'}>Postnummer</label>
               <input
-                className={
-                  !validations || !validations.zipCode ?
-                    'address-component-small-inputs form-control' :
-                    'address-component-small-inputs form-control validation-error'
-                }
+                className={classNames('address-component-small-inputs', 'form-control',
+                  {
+                    'validation-error': (validations ? validations.zipCode : false),
+                    'disabled': this.props.component.readOnly,
+                  })}
                 value={zipCode}
                 onChange={this.updateField.bind(null, AddressKeys.zipCode)}
                 onBlur={this.onBlurField.bind(null, AddressKeys.zipCode)}
+                disabled={this.props.component.readOnly}
               />
               {validations ?
                 renderValidationMessagesForComponent(validations[AddressKeys.zipCode], `${id}_${AddressKeys.zipCode}`)
-              : null}
+                : null}
             </div>
 
             <div className={'address-component-postplace'}>
               <label className={'address-component-label'}>Poststed</label>
               <input
-                className={'form-control'}
+                className={classNames('form-control', { 'disabled': this.props.component.readOnly })}
                 value={postPlace}
                 onChange={this.updateField.bind(null, AddressKeys.postPlace)}
                 onBlur={this.onBlurField.bind(null, AddressKeys.postPlace)}
+                disabled={this.props.component.readOnly}
               />
               {validations ?
                 renderValidationMessagesForComponent(validations[AddressKeys.postPlace],
                   `${id}_${AddressKeys.postPlace}`)
-              : null}
+                : null}
             </div>
           </div>
         </div>
@@ -247,54 +252,54 @@ export class AddressComponent extends React.Component<IAddressComponentProps, IA
             this.props.getTextResource(this.props.component.textResourceBindings.address)
         }</label>
         <input
-          className={'form-control'}
+          className={classNames('form-control', { 'disabled': this.props.component.readOnly })}
           value={address}
           onChange={this.updateField.bind(null, AddressKeys.address)}
           onBlur={this.onBlurField.bind(null, AddressKeys.address)}
         />
         {validations ?
-            renderValidationMessagesForComponent(validations[AddressKeys.address], `${id}_${AddressKeys.address}`)
+          renderValidationMessagesForComponent(validations[AddressKeys.address], `${id}_${AddressKeys.address}`)
           : null}
         <label className={'address-component-label'}>c/o eller annen tilleggsadresse</label>
         <input
-          className={'form-control'}
+          className={classNames('form-control', { 'disabled': this.props.component.readOnly })}
           value={careOf}
           onChange={this.updateField.bind(null, AddressKeys.careOf)}
           onBlur={this.onBlurField.bind(null, AddressKeys.careOf)}
         />
         {validations ?
-            renderValidationMessagesForComponent(validations[AddressKeys.careOf], `${id}_${AddressKeys.careOf}`)
+          renderValidationMessagesForComponent(validations[AddressKeys.careOf], `${id}_${AddressKeys.careOf}`)
           : null}
         <div className={'address-component-postplace-zipCode'}>
           <div className={'address-component-zipCode'}>
             <label className={'address-component-label'}>Postnummer</label>
             <br />
             <input
-              className={
-                !validations || !validations.zipCode ?
-                  'address-component-small-inputs form-control' :
-                  'address-component-small-inputs form-control validation-error'
-              }
+              className={classNames('address-component-small-inputs', 'form-control',
+                {
+                  'validation-error': (validations ? validations.zipCode : false),
+                  'disabled': this.props.component.readOnly,
+                })}
               value={zipCode}
               onChange={this.updateField.bind(null, AddressKeys.zipCode)}
               onBlur={this.onBlurField.bind(null, AddressKeys.zipCode)}
             />
             {validations ?
               renderValidationMessagesForComponent(validations[AddressKeys.zipCode], `${id}_${AddressKeys.zipCode}`)
-            : null}
+              : null}
           </div>
           <div className={'address-component-postplace'}>
             <label className={'address-component-label'}>Poststed</label>
             <br />
             <input
-              className={'form-control'}
+              className={classNames('form-control', { 'disabled': this.props.component.readOnly })}
               value={postPlace}
               onChange={this.updateField.bind(null, AddressKeys.postPlace)}
               onBlur={this.onBlurField.bind(null, AddressKeys.postPlace)}
             />
             {validations ?
               renderValidationMessagesForComponent(validations[AddressKeys.postPlace], `${id}_${AddressKeys.postPlace}`)
-            : null}
+              : null}
           </div>
         </div>
         <label className={'address-component-label'}>
@@ -308,18 +313,18 @@ export class AddressComponent extends React.Component<IAddressComponentProps, IA
           Den består av en bokstav og fire tall og skal være ført opp ved/på inngangsdøren din.
         </p>
         <input
-          className={
-            !validations.houseNumber ?
-              'address-component-small-inputs form-control' :
-              'address-component-small-inputs form-control validation-error'
-          }
+          className={classNames('address-component-small-inputs', 'form-control',
+            {
+              'validation-error': (validations ? validations.houseNumber : false),
+              'disabled': this.props.component.readOnly,
+            })}
           value={houseNumber}
           onChange={this.updateField.bind(null, AddressKeys.houseNumber)}
           onBlur={this.onBlurField.bind(null, AddressKeys.houseNumber)}
         />
         {validations ?
           renderValidationMessagesForComponent(validations[AddressKeys.houseNumber], `${id}_${AddressKeys.houseNumber}`)
-        : null}
+          : null}
       </div>
     );
   }
