@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createRef } from 'react';
-import '../../styles/InputComponent.css';
+import '../../styles/shared.css';
 
 export interface IDatePickerProps {
   id: string;
@@ -46,21 +46,22 @@ export class DatepickerComponent
   public render() {
     return (
       <div className='form-group a-form-group a-form-group-datepicker'>
-        <div className='input-group'>
+        <div className={'input-group' + (this.props.component.readOnly ? ' disabled' : '')}>
           <input
             type='text'
             id={this.props.id}
-            className={this.props.isValid ?
-              'form-control a-hasButton date' :
-              'form-control a-hasButton date validation-error'}
+            className={(this.props.component.readOnly ? 'disabled-date ' : '') +
+              (this.props.isValid ?
+                'form-control a-hasButton date' :
+                'form-control a-hasButton date validation-error')}
             onBlur={this.onDateChange}
             onChange={this.onDateChange}
-            disabled={this.props.component.disabled}
+            disabled={this.props.component.readOnly}
             required={this.props.component.required}
             value={this.state.value}
             ref={this.myDateCmp}
           />
-          <div className='input-group-prepend a-icon-right'>
+          <div className={'input-group-prepend a-icon-right' + (this.props.component.readOnly ? ' disabled-date' : '')}>
             <i className='ai ai-date' />
           </div>
         </div>
