@@ -1,37 +1,44 @@
-import { Checkbox, createStyles, WithStyles, withStyles } from '@material-ui/core';
+import { Checkbox, createMuiTheme, createStyles, WithStyles, withStyles } from '@material-ui/core';
 import React = require('react');
-
+import altinnTheme from '../theme/altinnStudioTheme';
 export interface IAltinnCheckBoxComponentProvidedProps extends WithStyles<typeof styles> {
+  checked: boolean;
+  disabled?: boolean;
   id?: any;
   onChangeFunction: any;
-  checked: boolean;
+  onKeyPressFunction?: any;
 }
 
 export interface IAltinnCheckBoxComponentState {
 }
+const theme = createMuiTheme(altinnTheme);
 
 const styles = () => createStyles({
   altinnCheckBox: {
-    paddingLeft: '0px',
-    paddingRight: '6px',
-    paddingTop: '6px',
+    'padding': '0px',
+    'color': theme.altinnPalette.primary.blueDark + ' !important',
+    '&span': {
+      height: '20px',
+      width: '20px',
+      marginRight: '6px',
+      marginTop: '6px',
+    },
   },
 });
 
-class AltinnCheckBox extends
-  React.Component<IAltinnCheckBoxComponentProvidedProps, IAltinnCheckBoxComponentState> {
-
-  public render() {
-    const { classes } = this.props;
-    return (
-      <Checkbox
-        id={this.props.id}
-        classes={{ root: classes.altinnCheckBox }}
-        checked={this.props.checked}
-        onChange={this.props.onChangeFunction}
-      />
-    );
-  }
-}
+const AltinnCheckBox = (props: IAltinnCheckBoxComponentProvidedProps) => {
+  const { classes } = props;
+  return (
+    <Checkbox
+      id={props.id}
+      className={classes.altinnCheckBox}
+      checked={props.checked}
+      onChange={props.onChangeFunction}
+      disabled={props.disabled ? props.disabled : false}
+      onKeyPress={props.onKeyPressFunction}
+      tabIndex={0}
+    />
+  );
+};
 
 export default withStyles(styles)(AltinnCheckBox);
